modules/MCQ/requirements.txt
streamlit>=1.36
supabase>=2.6
python-dotenv
